Major changes to PlotData wrapper
The PlotData wrapper is now able to select plot types, by input of a
dictionary of the DataFrame columns with the plot type for that column.

Plots will default to a scatter plot, unless the column name contains
“Model”, “Regression”, “-“, or “Fit” to represent either fitted data or
smoothed data in which case it comes out as a line plot.

The scatter marker size should now properly scale according to the
number of data points now.

A bug I can’t seem to fix is the colours of scatter plots. I can’t pin
the problem down exactly but the sometimes each scatter point is
different coloured.