Random changes
In NotebookTools, I have changed how the HDF5 databases are created
slightly: now it’s np.loadtxt instead of np.fromfile when generating
the arrays.

Large change in SpectralAnalysis is a new ScipyLinearRegression
function, which is able to perform a robust linear regression (treats
outliers with the l1_estimator) and bootstrap analysis in one function.
Use this from now on for a linear fit to data.